+++ client/modules/map/IndoorMapUtils.ts
