+++ client/modules/map/MapContext.tsx
@@ -12,7 +12,7 @@ import { getRoute } from './MapService';
 import { Location, Coordinates, Route, Level } from './Types';
 import { IndoorMap, indoorMaps } from './IndoorMap';
 import type { BBox } from 'geojson';
-import { bboxCenter, overlap } from './Utils';
+import { bboxCenter, overlap } from './IndoorMapUtils';
 import { default as turfDistance } from '@turf/distance';
 import { LocationSubscription, watchPositionAsync } from 'expo-location';
 import CoordinateService from '@/services/CoordinateService';
